pipex

+++ pipex/ft_strjoin.c
@@ -1,10 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strjoin.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: zkarapet <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/29 19:10:35 by zkarapet          #+#    #+#             */
+/*   Updated: 2022/07/01 19:19:14 by zkarapet         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "pipex.h"
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	unsigned long	i;
-	unsigned long	j;
-	char			*dst;
+	int		i;
+	int		j;
+	char	*dst;
 
 	i = 0;
 	j = 0;
